+++ apps/readest-app/src/components/settings/integrations/WebDAVForm.tsx
@@ -1,53 +1,24 @@
 import clsx from 'clsx';
-import React, { useEffect, useMemo, useState } from 'react';
-import {
-  MdFolder,
-  MdInsertDriveFile,
-  MdRefresh,
-  MdArrowBack,
-  MdVisibility,
-  MdVisibilityOff,
-  MdCloudSync,
-  MdDownload,
-  MdCheck,
-} from 'react-icons/md';
-import {
-  BsBook,
-  BsFiletypePdf,
-  BsFiletypeTxt,
-  BsFiletypeMd,
-  BsFiletypeXml,
-  BsFiletypePng,
-  BsFiletypeJpg,
-  BsFiletypeJson,
-  BsFiletypeOtf,
-  BsFiletypeTtf,
-  BsFiletypeWoff,
-} from 'react-icons/bs';
-import { LuBookImage } from 'react-icons/lu';
+import React, { useState } from 'react';
+import { MdVisibility, MdVisibilityOff, MdCloudSync } from 'react-icons/md';
 import { v4 as uuidv4 } from 'uuid';
 import { useEnv } from '@/context/EnvContext';
-import { useAuth } from '@/context/AuthContext';
 import { useTranslation } from '@/hooks/useTranslation';
 import { useSettingsStore } from '@/store/settingsStore';
 import { useLibraryStore } from '@/store/libraryStore';
 import { isTauriAppPlatform } from '@/services/environment';
 import { tauriDownload } from '@/utils/transfer';
 import { eventDispatcher } from '@/utils/event';
-import { ingestFile } from '@/services/ingestService';
 import {
   buildBasicAuthHeader,
   buildRequestUrl,
   checkConnection,
-  listDirectory,
   normalizeRootPath,
-  WebDAVEntry,
   WebDAVRequestError,
 } from '@/services/webdav/WebDAVClient';
 import { syncLibrary } from '@/services/webdav/WebDAVSync';
 import { buildWebDAVConnectSettings } from '@/services/webdav/webdavConnectSettings';
 import { getCoverFilename, getLocalBookFilename } from '@/utils/book';
-import { EXTS } from '@/libs/document';
 import {
   WEBDAV_SYNC_LOG_LIMIT,
   WebDAVSyncLogEntry,
@@ -58,11 +29,12 @@ import SubPageHeader from '../SubPageHeader';
 import {
   BoxedList,
   SectionTitle,
-  SettingLabel,
   SettingsRow,
   SettingsSwitchRow,
   SettingsSelect,
 } from '../primitives';
+import SyncHistoryPanel from './SyncHistoryPanel';
+import WebDAVBrowsePane from './WebDAVBrowsePane';
 
 interface WebDAVFormProps {
   onBack: () => void;
@@ -72,109 +44,41 @@ interface WebDAVFormProps {
  * WebDAV integration form. Two modes share the same panel:
  *
  * - Configuration: editable URL/username/password/root + Connect button.
- *   Lives in local state until Connect succeeds — only then do we persist
- *   the credentials via `saveSettings`. Failures surface via toast.
+ *   Lives in local state until Connect succeeds — only then do we
+ *   persist the credentials via `saveSettings`. Failures surface via
+ *   toast.
  *
- * - Browse: once configured, we render the contents of the root path the
- *   user picked. Sub-folders can be entered, and a breadcrumb / back-button
- *   lets the user navigate back to the saved root. Disconnect clears the
- *   stored config and returns to the configuration view.
+ * - Connected: renders the per-page sync controls (sub-toggles, Sync
+ *   now, sync-history) plus the {@link WebDAVBrowsePane} for the
+ *   stored root, and a Disconnect button. The browse pane is its own
+ *   component to keep this file legible — see its docstring.
  */
 const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
   const _ = useTranslation();
   const { settings, setSettings, saveSettings } = useSettingsStore();
   const { envConfig } = useEnv();
-  const { user } = useAuth();
 
   const stored = settings.webdav;
-  // Show the browse view only when an active connection is configured. We
-  // rely on `enabled` (set by Connect, cleared by Disconnect) rather than
-  // looking at serverUrl/username, so Disconnect always returns the user to
-  // the configuration form even if we keep their previous URL pre-filled.
+  // Show the browse view only when an active connection is configured.
+  // We rely on `enabled` (set by Connect, cleared by Disconnect) rather
+  // than looking at serverUrl/username so Disconnect always returns the
+  // user to the configuration form even if we keep their previous URL
+  // pre-filled.
   const isConfigured = !!stored?.enabled && !!stored?.serverUrl;
 
-  // Editable form state — initialised from saved settings so re-entering the
-  // sub-page after a previous configure preserves what the user typed.
+  // Editable form state — initialised from saved settings so re-entering
+  // the sub-page after a previous configure preserves what the user
+  // typed.
   const [url, setUrl] = useState(stored?.serverUrl || '');
   const [username, setUsername] = useState(stored?.username || '');
   const [password, setPassword] = useState(stored?.password || '');
   const [rootPath, setRootPath] = useState(stored?.rootPath || '/');
   const [isConnecting, setIsConnecting] = useState(false);
   const [showPassword, setShowPassword] = useState(false);
-  // Library-wide Sync now state — surfaces a progress hint while we walk
-  // through the bookshelf and disables the button to prevent re-entry.
+  // Library-wide Sync now state — surfaces a progress hint while we
+  // walk through the bookshelf and disables the button to prevent
+  // re-entry.
   const [syncProgressLabel, setSyncProgressLabel] = useState<string | null>(null);
-
-  // Browse-mode state. `currentPath` may differ from the stored rootPath
-  // once the user drills into sub-folders.
-  const [currentPath, setCurrentPath] = useState<string>(stored?.rootPath || '/');
-  const [entries, setEntries] = useState<WebDAVEntry[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [loadError, setLoadError] = useState<string | null>(null);
-  // Increments on Refresh — used purely as an effect dependency so the
-  // listing reloads even when `currentPath` hasn't changed.
-  const [reloadTick, setReloadTick] = useState(0);
-  // Per-entry download status keyed by remote path. Resets when the user
-  // navigates or refreshes — within a session, a successful download is
-  // marked "done" so the button stops inviting a redundant re-tap.
-  // `importBook`'s hash dedupe protects us if the user does click again,
-  // but a quiet visual cue is friendlier than a no-op.
-  const [downloadStatus, setDownloadStatus] = useState<
-    Record<string, 'downloading' | 'done' | 'error'>
-  >({});
-
-  const savedRoot = useMemo(() => normalizeRootPath(stored?.rootPath || '/'), [stored?.rootPath]);
-
-  // When we transition into browse-mode (right after Connect, or when the
-  // user re-opens the page already configured), pull the listing.
-  useEffect(() => {
-    if (!isConfigured) return;
-    setCurrentPath((prev) => prev || savedRoot);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isConfigured]);
-
-  useEffect(() => {
-    if (!isConfigured || !currentPath) return;
-    let cancelled = false;
-    // Reset per-entry download status whenever we (re)load a directory:
-    // stale "done" badges from a previous folder would otherwise confuse
-    // users who navigate back to that folder.
-    setDownloadStatus({});
-    const load = async () => {
-      setIsLoading(true);
-      setLoadError(null);
-      try {
-        const list = await listDirectory(
-          {
-            serverUrl: stored.serverUrl,
-            username: stored.username,
-            password: stored.password,
-          },
-          currentPath,
-        );
-        if (!cancelled) setEntries(list);
-      } catch (e) {
-        if (!cancelled) {
-          setEntries([]);
-          setLoadError((e as Error).message || _('Failed to load directory'));
-        }
-      } finally {
-        if (!cancelled) setIsLoading(false);
-      }
-    };
-    load();
-    return () => {
-      cancelled = true;
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [
-    isConfigured,
-    currentPath,
-    reloadTick,
-    stored?.serverUrl,
-    stored?.username,
-    stored?.password,
-  ]);
 
   const handleConnect = async () => {
     if (!url || !username) return;
@@ -191,8 +95,8 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
     }
     // Spread previous webdav state so a reconnect preserves bookkeeping
     // fields earned by prior use — deviceId, syncBooks, strategy,
-    // syncProgress, syncNotes, lastSyncedAt, syncLog. Rotating deviceId on
-    // reconnect would make this device look new to the cross-device
+    // syncProgress, syncNotes, lastSyncedAt, syncLog. Rotating deviceId
+    // on reconnect would make this device look new to the cross-device
     // clobber check in `RemoteBookConfig.writerDeviceId`.
     const newSettings = {
       ...settings,
@@ -205,7 +109,6 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
     };
     setSettings(newSettings);
     await saveSettings(envConfig, newSettings);
-    setCurrentPath(normalizedRoot);
     setIsConnecting(false);
     eventDispatcher.dispatch('toast', { type: 'info', message: _('Connected') });
   };
@@ -220,24 +123,17 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
     };
     setSettings(newSettings);
     await saveSettings(envConfig, newSettings);
-    // Reset browse-mode bookkeeping so re-entering Connect doesn't briefly
-    // flash a stale 401 error or the previous directory listing.
-    setEntries([]);
-    setLoadError(null);
-    setIsLoading(false);
-    // Keep the password pre-filled (masked) so the user can reconnect with
-    // a single click — they can still toggle visibility via the eye icon.
+    // Keep the password pre-filled (masked) so the user can reconnect
+    // with a single click — they can still toggle visibility via the
+    // eye icon.
     setShowPassword(false);
     eventDispatcher.dispatch('toast', { type: 'info', message: _('Disconnected') });
   };
 
-  const handleEntryClick = (entry: WebDAVEntry) => {
-    if (entry.isDirectory) setCurrentPath(entry.path);
-  };
-
   // —— Sync sub-toggles & manual triggers ——
-  // The toggles persist via saveSettings synchronously (debouncing isn't
-  // worth the extra state — users tap each toggle at most once per session).
+  // The toggles persist via saveSettings synchronously (debouncing
+  // isn't worth the extra state — users tap each toggle at most once
+  // per session).
   //
   // IMPORTANT: read latest settings from the store (NOT the closure
   // variable) when computing `next`. `handleSyncNow` issues two
@@ -269,8 +165,8 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
    *
    * We deliberately re-read settings from the store at write time
    * (rather than closing over `stored`) so concurrent updates — e.g.
-   * the user flips the syncBooks toggle while a Sync now is in flight —
-   * don't clobber each other. The 10-entry cap matches
+   * the user flips the syncBooks toggle while a Sync now is in flight
+   * — don't clobber each other. The 10-entry cap matches
    * `WEBDAV_SYNC_LOG_LIMIT` and trims oldest-first; we keep the
    * persisted JSON small so settings.json round-trips on every app
    * start stay cheap.
@@ -289,8 +185,9 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
 
   /**
    * Manual "Sync now" — push every book in the local library up to the
-   * remote in a single sequential pass. We don't pull here; the per-book
-   * Reader hook handles incoming changes when the user opens a book.
+   * remote in a single sequential pass. We don't pull here; the per-
+   * book Reader hook handles incoming changes when the user opens a
+   * book.
    *
    * Why sequential: shared WebDAV servers (NextCloud, Synology, …) are
    * not happy with parallel PUTs from one user, and a steady linear
@@ -313,10 +210,11 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
 
     const eligibleBooks = currentLibrary.filter((b) => !b.deletedAt);
 
-    // Lazily ensure a deviceId so the first cross-device sync attributes
-    // its rows correctly. The same field is also touched by the Reader
-    // hook on first push; doing it here too keeps the Sync now path
-    // self-sufficient when the user has never opened a book yet.
+    // Lazily ensure a deviceId so the first cross-device sync
+    // attributes its rows correctly. The same field is also touched by
+    // the Reader hook on first push; doing it here too keeps the Sync
+    // now path self-sufficient when the user has never opened a book
+    // yet.
     let deviceId = stored.deviceId;
     if (!deviceId) {
       deviceId = uuidv4();
@@ -409,8 +307,8 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
             book.coverImageUrl = await appService.generateCoverImageUrl(book);
           } catch (e) {
             // Missing or broken cover shouldn't block adding the book —
-            // the bookshelf renders a placeholder when coverImageUrl is
-            // empty.
+            // the bookshelf renders a placeholder when coverImageUrl
+            // is empty.
             console.warn('WD library sync: cover URL generation failed', book.hash, e);
             book.coverImageUrl = null;
           }
@@ -529,9 +427,9 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
       eventDispatcher.dispatch('toast', { type: 'error', message });
       // Persist a "failure" entry so the user can show what went wrong
       // without rummaging through the dev console. We don't have a
-      // SyncLibraryResult to draw counters from (the run aborted before
-      // returning), so all the count fields stay zero except totalBooks
-      // for context.
+      // SyncLibraryResult to draw counters from (the run aborted
+      // before returning), so all the count fields stay zero except
+      // totalBooks for context.
       const entry: WebDAVSyncLogEntry = {
         id: uuidv4(),
         startedAt,
@@ -555,122 +453,9 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
     }
   };
 
-  const handleNavigateUp = () => {
-    if (currentPath === savedRoot) return;
-    const trimmed = currentPath.replace(/\/+$/, '');
-    const idx = trimmed.lastIndexOf('/');
-    const parent = idx <= 0 ? '/' : trimmed.slice(0, idx);
-    // Don't escape above the saved root — the integration is scoped to it.
-    if (!parent.startsWith(savedRoot)) {
-      setCurrentPath(savedRoot);
-    } else {
-      setCurrentPath(parent);
-    }
-  };
-
-  const handleRefresh = () => {
-    setReloadTick((n) => n + 1);
-  };
-
-  /**
-   * Download a single remote file into the local library.
-   *
-   * Mirrors the OPDS auto-download path: stream the bytes to a Cache
-   * file via `tauriDownload` (avoids the WebView<->Tauri IPC binder
-   * limit that kills mid-write transfers on Android), then hand the
-   * file to {@link ingestFile} — the channel-agnostic importer used
-   * by every capture path (local folder import, Send-to-Readest,
-   * inbox drainer). It extracts metadata, writes the cover, computes
-   * the content hash (so re-downloading the same book is a deduped
-   * no-op), and produces a Book entry; honouring the user's autoUpload
-   * setting means a logged-in user with autoUpload on will see the
-   * downloaded book pushed to the readest cloud too — same behaviour
-   * as dragging the file in by hand.
-   *
-   * Web/desktop builds without `tauriDownload` aren't expected to
-   * exercise this path — the Settings page is gated to Tauri platforms.
-   * If a web build ever reaches here, we surface a clear toast instead
-   * of silently doing nothing.
-   */
-  const handleDownloadEntry = async (entry: WebDAVEntry) => {
-    if (entry.isDirectory) return;
-    if (!isSupportedBookExt(entry.name)) return;
-    if (downloadStatus[entry.path] === 'downloading' || downloadStatus[entry.path] === 'done') {
-      return;
-    }
-    if (!isTauriAppPlatform()) {
-      eventDispatcher.dispatch('toast', {
-        type: 'error',
-        message: _('File download is only supported on the desktop and mobile apps.'),
-      });
-      return;
-    }
-    const appService = await envConfig.getAppService();
-    if (!appService) return;
-
-    setDownloadStatus((prev) => ({ ...prev, [entry.path]: 'downloading' }));
-    try {
-      // Stream into Cache under a unique filename so a parallel download
-      // (or a re-tap before the previous one finished) doesn't clobber
-      // the in-flight bytes. We don't bother with sanitisation here —
-      // the suffix dominates collision probability.
-      const safeName = entry.name.replaceAll(/[/\\:*?"<>|]/g, '_').slice(0, 200) || 'download';
-      const cacheName = `webdav-${Date.now()}-${safeName}`;
-      const dst = await appService.resolveFilePath(cacheName, 'Cache');
-      const url = buildRequestUrl(stored.serverUrl, entry.path);
-      const headers = {
-        Authorization: buildBasicAuthHeader(stored.username, stored.password),
-      };
-      await tauriDownload(url, dst, undefined, headers);
-
-      // Run import against a fresh library snapshot — the user may have
-      // imported books elsewhere since this page mounted. ingestFile
-      // delegates to importBook which mutates the array in place; we
-      // must persist + push it back into the store afterwards for the
-      // bookshelf to reflect the new entry.
-      const { library: storeLibrary, libraryLoaded, setLibrary } = useLibraryStore.getState();
-      const library = libraryLoaded ? [...storeLibrary] : await appService.loadLibraryBooks();
-      const imported = await ingestFile(
-        { file: dst, books: library },
-        { appService, settings, isLoggedIn: !!user },
-      );
-      // Best-effort cleanup of the cache file. ingestFile -> importBook
-      // copies the bytes into the per-hash Books folder, so this temp
-      // is no longer needed regardless of success/failure.
-      try {
-        await appService.deleteFile(dst, 'None');
-      } catch {
-        // Cache deletion is non-critical — leave it for the OS to GC.
-      }
-      if (!imported) {
-        throw new Error('Import returned null');
-      }
-      await appService.saveLibraryBooks(library);
-      setLibrary(library);
-
-      setDownloadStatus((prev) => ({ ...prev, [entry.path]: 'done' }));
-      eventDispatcher.dispatch('toast', {
-        type: 'info',
-        message: _('Downloaded "{{title}}" to your library.', {
-          title: imported.title || entry.name,
-        }),
-      });
-    } catch (e) {
-      console.warn('WebDAV download failed', entry.path, e);
-      setDownloadStatus((prev) => ({ ...prev, [entry.path]: 'error' }));
-      eventDispatcher.dispatch('toast', {
-        type: 'error',
-        message: _('Failed to download "{{name}}": {{error}}', {
-          name: entry.name,
-          error: (e as Error).message ?? String(e),
-        }),
-      });
-    }
-  };
-
   const description: string = isConfigured
     ? _('Browsing {{path}} on {{server}}', {
-        path: savedRoot,
+        path: normalizeRootPath(stored.rootPath || '/'),
         server: stored.serverUrl,
       })
     : _('Connect to a WebDAV server to browse your remote files.');
@@ -686,9 +471,9 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
 
       {isConfigured ? (
         <div className='space-y-5'>
-          {/* Sync controls — sub-category toggles, conflict strategy, and
-              a manual "Sync now" button. Mirrors the layout used by
-              KOSyncForm so users get a consistent surface. */}
+          {/* Sync controls — sub-category toggles, conflict strategy,
+              and a manual "Sync now" button. Mirrors the layout used
+              by KOSyncForm so users get a consistent surface. */}
           <BoxedList>
             <SettingsSwitchRow
               label={_('Upload Book Files')}
@@ -753,163 +538,7 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
               lives before their first run. */}
           <SyncHistoryPanel entries={stored.syncLog ?? []} onClear={handleClearSyncLog} t={_} />
 
-          <div className='flex items-center justify-between gap-3 px-1'>
-            <div className='flex min-w-0 items-center gap-2'>
-              <button
-                type='button'
-                onClick={handleNavigateUp}
-                disabled={currentPath === savedRoot}
-                className={clsx(
-                  'btn btn-ghost btn-sm h-8 min-h-8 gap-1 px-2',
-                  currentPath === savedRoot && 'opacity-40',
-                )}
-                title={_('Up')}
-                aria-label={_('Up')}
-              >
-                <MdArrowBack className='h-4 w-4' />
-              </button>
-              <span className='truncate text-sm' title={currentPath}>
-                {currentPath}
-              </span>
-            </div>
-            <button
-              type='button'
-              onClick={handleRefresh}
-              className='btn btn-ghost btn-sm h-8 min-h-8 px-2'
-              title={_('Refresh')}
-              aria-label={_('Refresh')}
-            >
-              <MdRefresh className='h-4 w-4' />
-            </button>
-          </div>
-
-          <div className='card eink-bordered border-base-200 bg-base-100 overflow-hidden border'>
-            {isLoading ? (
-              <div className='flex min-h-32 items-center justify-center py-8'>
-                <span className='loading loading-spinner loading-md' />
-              </div>
-            ) : loadError ? (
-              <div className='text-error px-4 py-6 text-center text-sm'>{loadError}</div>
-            ) : entries.length === 0 ? (
-              <div className='text-base-content/60 px-4 py-6 text-center text-sm'>
-                {_('Empty directory')}
-              </div>
-            ) : (
-              <ul className='divide-base-200 divide-y'>
-                {entries.map((entry) => {
-                  const FileIcon = entry.isDirectory ? MdFolder : getEntryIcon(entry.name);
-                  const canDownload = !entry.isDirectory && isSupportedBookExt(entry.name);
-                  const dlState = downloadStatus[entry.path];
-                  // The whole row is interactive only for directories; for
-                  // files, only the trailing download button (when shown)
-                  // is clickable. We use a div+role=button on the row so
-                  // the inner download button isn't nested inside another
-                  // <button> (invalid HTML, click bubbling).
-                  const rowClickable = entry.isDirectory;
-                  return (
-                    <li key={entry.path}>
-                      <div
-                        role={rowClickable ? 'button' : undefined}
-                        tabIndex={rowClickable ? 0 : -1}
-                        onClick={rowClickable ? () => handleEntryClick(entry) : undefined}
-                        onKeyDown={
-                          rowClickable
-                            ? (e) => {
-                                if (e.key === 'Enter' || e.key === ' ') {
-                                  e.preventDefault();
-                                  handleEntryClick(entry);
-                                }
-                              }
-                            : undefined
-                        }
-                        className={clsx(
-                          'group flex w-full items-center gap-3 px-4 py-3 text-left',
-                          'transition-colors duration-150',
-                          rowClickable ? 'hover:bg-base-200/60 cursor-pointer' : 'cursor-default',
-                        )}
-                      >
-                        <span
-                          className={clsx(
-                            'flex h-8 w-8 flex-shrink-0 items-center justify-center rounded',
-                            'bg-base-200 text-base-content/70',
-                          )}
-                        >
-                          <FileIcon className='h-4 w-4' />
-                        </span>
-                        <div className='flex min-w-0 flex-1 flex-col gap-0.5'>
-                          {/* line-clamp-none cancels SettingLabel's default
-                              line-clamp-2 so long file names wrap freely;
-                              break-all + whitespace-normal handle hyphenless
-                              unicode names that don't have a wrap opportunity. */}
-                          <SettingLabel className='line-clamp-none whitespace-normal break-all'>
-                            {entry.name}
-                          </SettingLabel>
-                          {/* Secondary metadata line. Files show size +
-                              mtime; directories only show mtime (most
-                              servers don't report aggregated child sizes
-                              on a collection). The whole line is gated
-                              on at least one field being available so we
-                              don't render an empty span. */}
-                          {((!entry.isDirectory && typeof entry.size === 'number') ||
-                            entry.lastModified) && (
-                            <span className='text-base-content/60 flex flex-wrap gap-x-2 text-[0.75em]'>
-                              {!entry.isDirectory && typeof entry.size === 'number' && (
-                                <span>{formatSize(entry.size)}</span>
-                              )}
-                              {entry.lastModified && (
-                                <span title={entry.lastModified}>
-                                  {formatLastModified(entry.lastModified)}
-                                </span>
-                              )}
-                            </span>
-                          )}
-                        </div>
-                        {canDownload && (
-                          <button
-                            type='button'
-                            onClick={(e) => {
-                              // Stop propagation defensively — the parent
-                              // div is non-clickable for files today, but
-                              // keeps us safe if that ever changes.
-                              e.stopPropagation();
-                              handleDownloadEntry(entry);
-                            }}
-                            disabled={dlState === 'downloading' || dlState === 'done'}
-                            className={clsx(
-                              'btn btn-ghost btn-sm h-8 min-h-8 flex-shrink-0 px-2',
-                              (dlState === 'downloading' || dlState === 'done') && 'opacity-60',
-                            )}
-                            title={
-                              dlState === 'done'
-                                ? _('Already downloaded in this session')
-                                : dlState === 'downloading'
-                                  ? _('Downloading…')
-                                  : _('Download to library')
-                            }
-                            aria-label={
-                              dlState === 'done'
-                                ? _('Already downloaded in this session')
-                                : dlState === 'downloading'
-                                  ? _('Downloading…')
-                                  : _('Download to library')
-                            }
-                          >
-                            {dlState === 'downloading' ? (
-                              <span className='loading loading-spinner loading-xs' />
-                            ) : dlState === 'done' ? (
-                              <MdCheck className='h-4 w-4' />
-                            ) : (
-                              <MdDownload className='h-4 w-4' />
-                            )}
-                          </button>
-                        )}
-                      </div>
-                    </li>
-                  );
-                })}
-              </ul>
-            )}
-          </div>
+          <WebDAVBrowsePane settings={stored} t={_} />
 
           <div className='flex justify-end'>
             <button
@@ -1038,408 +667,6 @@ const WebDAVForm: React.FC<WebDAVFormProps> = ({ onBack }) => {
               </button>
             </div>
           </form>
-        </div>
-      )}
-    </div>
-  );
-};
-
-const formatSize = (bytes: number): string => {
-  if (!Number.isFinite(bytes) || bytes < 0) return '';
-  const units = ['B', 'KB', 'MB', 'GB', 'TB'];
-  let value = bytes;
-  let unit = 0;
-  while (value >= 1024 && unit < units.length - 1) {
-    value /= 1024;
-    unit += 1;
-  }
-  const formatted =
-    unit === 0 ? value.toFixed(0) : value.toFixed(value >= 100 ? 0 : value >= 10 ? 1 : 2);
-  return `${formatted} ${units[unit]}`;
-};
-
-/**
- * Render the WebDAV-supplied last-modified timestamp in a compact, locale-
- * aware form for the secondary line under each file. Servers usually emit
- * RFC 1123 ("Wed, 02 Oct 2002 13:00:00 GMT") via `getlastmodified`, but a
- * handful return ISO-8601 — `Date` parses both. If the value can't be
- * understood we return an empty string so the row simply omits the field
- * rather than rendering "Invalid Date".
- *
- * The full timestamp is exposed as a `title` on the wrapping span so power
- * users can hover for the precise time without us spending pixels on it.
- */
-const formatLastModified = (raw: string): string => {
-  const ts = Date.parse(raw);
-  if (!Number.isFinite(ts)) return '';
-  try {
-    return new Date(ts).toLocaleString(undefined, {
-      year: 'numeric',
-      month: 'short',
-      day: 'numeric',
-      hour: '2-digit',
-      minute: '2-digit',
-    });
-  } catch {
-    // Some embedded WebViews (older Android) reject options bags with
-    // both date and time fields — fall back to the default formatter
-    // rather than showing nothing.
-    return new Date(ts).toLocaleString();
-  }
-};
-
-/**
- * True when the filename's extension matches one of the reader-supported
- * book formats declared in `libs/document.ts`. Used to gate the per-entry
- * download button: only files we can actually open get the affordance, so
- * the user doesn't waste a tap pulling down something the library can't
- * import (font files, README.txt, hidden dotfiles, …).
- *
- * Keep aligned with `EXTS` — adding a new format there should automatically
- * light up the download button for that extension.
- */
-const SUPPORTED_BOOK_EXTS = new Set<string>(Object.values(EXTS).map((e) => e.toLowerCase()));
-
-const isSupportedBookExt = (filename: string): boolean => {
-  const m = filename.match(/\.([^.]+)$/);
-  const ext = m && m[1] ? m[1].toLowerCase() : '';
-  return !!ext && SUPPORTED_BOOK_EXTS.has(ext);
-};
-
-/**
- * Pick a per-file icon based on the entry's extension. Reader-recognised
- * formats get a specific icon (`BsFiletypePdf`, `BsFiletypeTxt`,
- * `BsFiletypeMd` are pixel-perfect matches; EPUB / MOBI / AZW / FB2 fall
- * back to the generic `BsBook`; CBZ uses the comic-friendly `LuBookImage`),
- * everything else stays on the neutral `MdInsertDriveFile`. Cover images
- * and the library/config JSON files that live next to each book also get
- * their own icons so the readest-managed subtree is self-explanatory.
- *
- * Keep in sync with `EXTS` in `libs/document.ts` when a new format is added.
- */
-const getEntryIcon = (filename: string): React.ComponentType<{ className?: string }> => {
-  const m = filename.match(/\.([^.]+)$/);
-  const ext = m && m[1] ? m[1].toLowerCase() : '';
-  switch (ext) {
-    case 'pdf':
-      return BsFiletypePdf;
-    case 'txt':
-      return BsFiletypeTxt;
-    case 'md':
-      return BsFiletypeMd;
-    case 'fb2':
-    case 'fbz':
-      return BsFiletypeXml;
-    case 'cbz':
-      return LuBookImage;
-    case 'epub':
-    case 'mobi':
-    case 'azw':
-    case 'azw3':
-      return BsBook;
-    case 'png':
-      return BsFiletypePng;
-    case 'jpg':
-    case 'jpeg':
-      return BsFiletypeJpg;
-    case 'json':
-      return BsFiletypeJson;
-    case 'xml':
-      return BsFiletypeXml;
-    case 'otf':
-      return BsFiletypeOtf;
-    case 'ttf':
-      return BsFiletypeTtf;
-    case 'woff':
-    case 'woff2':
-      return BsFiletypeWoff;
-    default:
-      return MdInsertDriveFile;
-  }
-};
-
-/**
- * Diagnostic surface for the most-recent ten Sync now runs.
- *
- * Why a separate component (rather than inline JSX in WebDAVForm):
- *  - Lets the outer form file stay legible at ~1000 lines; the panel
- *    has its own state model (which entry is expanded) that doesn't
- *    belong in the parent.
- *  - Keeps the rendering of an entry — counters, failure list,
- *    duration — colocated with the component that owns it. The parent
- *    only knows about "the log" as a whole and how to clear it.
- *
- * The component is presentational: all persistence happens in the
- * parent (`appendSyncLogEntry` / `handleClearSyncLog`). We accept the
- * translation function `t` rather than calling `useTranslation` here
- * so the parent can keep a single source of locale truth.
- */
-interface SyncHistoryPanelProps {
-  entries: WebDAVSyncLogEntry[];
-  onClear: () => void | Promise<void>;
-  t: (key: string, params?: Record<string, string | number>) => string;
-}
-
-const SyncHistoryPanel: React.FC<SyncHistoryPanelProps> = ({ entries, onClear, t }) => {
-  // Only one entry expanded at a time keeps the panel scannable on
-  // mobile — multiple open rows can quickly push the disconnect button
-  // off-screen. Set to null when no row is expanded.
-  const [expandedId, setExpandedId] = useState<string | null>(null);
-
-  const hasEntries = entries.length > 0;
-
-  return (
-    <BoxedList>
-      <SettingsRow
-        label={t('Sync History')}
-        description={t("Manual syncs only — automatic syncs while reading aren't logged here.")}
-      >
-        {hasEntries ? (
-          <button
-            type='button'
-            onClick={() => onClear()}
-            className='btn btn-ghost btn-sm h-8 min-h-8 px-2'
-            title={t('Clear Sync History')}
-            aria-label={t('Clear Sync History')}
-          >
-            {t('Clear')}
-          </button>
-        ) : (
-          <span className='text-base-content/50 text-xs'>{t('No manual syncs yet')}</span>
-        )}
-      </SettingsRow>
-      {hasEntries && (
-        <ul className='divide-base-200 divide-y'>
-          {entries.map((entry) => {
-            const isExpanded = expandedId === entry.id;
-            return (
-              <li key={entry.id} className='px-4 py-3'>
-                <button
-                  type='button'
-                  onClick={() => setExpandedId(isExpanded ? null : entry.id)}
-                  className='group flex w-full items-center gap-3 text-left'
-                  aria-expanded={isExpanded}
-                >
-                  <SyncStatusBadge status={entry.status} t={t} />
-                  <div className='flex min-w-0 flex-1 flex-col gap-0.5'>
-                    <span className='text-sm'>{formatSyncSummaryLine(entry, t)}</span>
-                    <span className='text-base-content/60 text-[0.75em]'>
-                      {formatSyncTimestamp(entry.startedAt, entry.finishedAt, t)}
-                    </span>
-                  </div>
-                  <span
-                    className={clsx(
-                      'text-base-content/50 transition-transform',
-                      isExpanded && 'rotate-90',
-                    )}
-                    aria-hidden
-                  >
-                    ›
-                  </span>
-                </button>
-                {isExpanded && <SyncHistoryDetails entry={entry} t={t} />}
-              </li>
-            );
-          })}
-        </ul>
-      )}
-    </BoxedList>
-  );
-};
-
-/**
- * Coloured pill summarising an entry's status. We pick semantic
- * Tailwind utilities (success / warning / error) so the badge respects
- * the user's theme (eink, dark, light) without per-mode overrides.
- */
-const SyncStatusBadge: React.FC<{ status: WebDAVSyncLogStatus; t: SyncHistoryPanelProps['t'] }> = ({
-  status,
-  t,
-}) => {
-  const map: Record<WebDAVSyncLogStatus, { label: string; className: string }> = {
-    success: { label: t('OK'), className: 'bg-success/15 text-success' },
-    partial: { label: t('Partial'), className: 'bg-warning/15 text-warning' },
-    failure: { label: t('Failed'), className: 'bg-error/15 text-error' },
-  };
-  const { label, className } = map[status];
-  return (
-    <span
-      className={clsx(
-        'flex h-6 flex-shrink-0 items-center rounded px-2 text-[0.7rem] font-medium',
-        className,
-      )}
-    >
-      {label}
-    </span>
-  );
-};
-
-/**
- * Build the one-line summary shown next to each history row's status
- * badge. We re-derive it from the structured counters (rather than
- * reusing the toast's `entry.summary`) so the text in the log stays
- * compact even when the original toast was multi-line.
- */
-const formatSyncSummaryLine = (
-  entry: WebDAVSyncLogEntry,
-  t: SyncHistoryPanelProps['t'],
-): string => {
-  if (entry.status === 'failure') {
-    return entry.errorMessage || t('Sync failed');
-  }
-  const parts: string[] = [];
-  if (entry.booksDownloaded > 0) {
-    parts.push(t('{{n}} downloaded', { n: entry.booksDownloaded }));
-  }
-  if (entry.filesUploaded > 0) {
-    parts.push(t('{{n}} uploaded', { n: entry.filesUploaded }));
-  }
-  if (entry.configsUploaded > 0 || entry.configsDownloaded > 0) {
-    parts.push(t('{{n}} progress', { n: entry.configsUploaded + entry.configsDownloaded }));
-  }
-  if (entry.failures > 0) {
-    parts.push(t('{{n}} failed', { n: entry.failures }));
-  }
-  return parts.length > 0 ? parts.join(' · ') : t('Up to date');
-};
-
-/**
- * "Mar 18, 14:32 · 4.2 s" — short locale-aware timestamp plus a
- * duration so users can spot abnormally slow runs at a glance.
- */
-const formatSyncTimestamp = (
-  startedAt: number,
-  finishedAt: number,
-  t: SyncHistoryPanelProps['t'],
-): string => {
-  const when = new Date(startedAt).toLocaleString(undefined, {
-    month: 'short',
-    day: 'numeric',
-    hour: '2-digit',
-    minute: '2-digit',
-  });
-  const durMs = Math.max(0, finishedAt - startedAt);
-  const dur = durMs >= 1000 ? `${(durMs / 1000).toFixed(1)} s` : `${durMs} ms`;
-  return t('{{when}} · {{dur}}', { when, dur });
-};
-
-/**
- * Expanded body of one history entry: the full counter grid plus the
- * per-book failure list when present. Counters that are zero are
- * suppressed so the grid only shows what actually happened — a
- * partial-failure row with three items is much easier to read than
- * the same row with seven zeroes interleaved.
- */
-const SyncHistoryDetails: React.FC<{
-  entry: WebDAVSyncLogEntry;
-  t: SyncHistoryPanelProps['t'];
-}> = ({ entry, t }) => {
-  // Counters are grouped semantically so the user can scan them at a
-  // glance instead of treating eight numbers as a flat blob:
-  //   - "activity": work performed during this run
-  //   - "skipped":  work that was deduped / no-op'd
-  //   - "outcome":  totals + failure count for at-a-glance triage
-  // Each group renders independently and is separated by a divider so
-  // it's visually obvious that "Configs uploaded" and "Total books"
-  // are different things — they previously sat side-by-side in a
-  // single grid which read as one block.
-  const groups: { label: string; value: number }[][] = [
-    [
-      { label: t('Books downloaded'), value: entry.booksDownloaded },
-      { label: t('Files uploaded'), value: entry.filesUploaded },
-      { label: t('Configs uploaded'), value: entry.configsUploaded },
-      { label: t('Configs downloaded'), value: entry.configsDownloaded },
-      { label: t('Covers uploaded'), value: entry.coversUploaded },
-    ],
-    [{ label: t('Files in sync'), value: entry.filesAlreadyInSync }],
-    [
-      { label: t('Failures'), value: entry.failures },
-      { label: t('Total books'), value: entry.totalBooks },
-    ],
-  ]
-    // Suppress zero-only groups entirely so we don't render an empty
-    // section + divider for a group whose every counter happens to be
-    // zero this run (common: 'skipped' and 'outcome' rows on a quiet
-    // sync). The within-group filter keeps individual zero entries out
-    // of mixed groups.
-    .map((group) => group.filter((c) => c.value > 0))
-    .filter((group) => group.length > 0);
-
-  return (
-    <div className='mt-3 flex flex-col gap-3 pl-9'>
-      {groups.length > 0 && (
-        // Six-column grid: each of the three semantic groups occupies
-        // a (label-column, value-column) pair. Label columns flex with
-        // available space and wrap naturally for long strings like
-        // "Configs uploaded"; value columns are sized to content so
-        // the numbers stay tightly packed against their labels. Border
-        // dividers between every other column visually separate the
-        // three groups; we draw them with `border-l` on columns 3 and
-        // 5 rather than CSS `divide-x` because divide-x can't honour
-        // the "skip every two columns" pattern.
-        <div
-          className={clsx(
-            'border-base-200 grid rounded border',
-            'gap-x-3 gap-y-2 px-3 py-2 text-xs',
-          )}
-          style={{
-            gridTemplateColumns: 'minmax(0, 1fr) auto minmax(0, 1fr) auto minmax(0, 1fr) auto',
-          }}
-        >
-          {(() => {
-            // All three columns share row count to keep the grid rows
-            // aligned. Compute it once outside the per-group map so
-            // each column sees the same value.
-            const maxRows = Math.max(...groups.map((g) => g.length), 0);
-            return [0, 1, 2].map((groupIdx) => {
-              const group = groups[groupIdx] ?? [];
-              const cells: React.ReactNode[] = [];
-              for (let row = 0; row < maxRows; row++) {
-                const c = group[row];
-                cells.push(
-                  <div
-                    key={`l-${groupIdx}-${row}`}
-                    className={clsx(
-                      'text-base-content/60 leading-tight',
-                      // Group separator: every group except the first
-                      // gets a left border on its label column. The
-                      // negative left margin offsets the gap so the
-                      // line falls inside the gutter rather than
-                      // beside the text itself.
-                      groupIdx > 0 && 'border-base-200 -ml-3 border-l pl-3',
-                    )}
-                  >
-                    {c?.label ?? ''}
-                  </div>,
-                );
-                cells.push(
-                  <div key={`v-${groupIdx}-${row}`} className='text-end font-medium tabular-nums'>
-                    {c?.value ?? ''}
-                  </div>,
-                );
-              }
-              return cells;
-            });
-          })()}
-        </div>
-      )}
-      {entry.errorMessage && (
-        <div className='text-error/90 break-words text-xs'>
-          <span className='text-base-content/60 mr-1'>{t('Error:')}</span>
-          {entry.errorMessage}
-        </div>
-      )}
-      {entry.failedBooks && entry.failedBooks.length > 0 && (
-        <div className='flex flex-col gap-1'>
-          <span className='text-base-content/60 text-xs'>{t('Failed books')}</span>
-          <ul className='flex flex-col gap-1 text-xs'>
-            {entry.failedBooks.map((f) => (
-              <li key={f.hash} className='border-base-200 break-words rounded border px-2 py-1.5'>
-                <div className='font-medium'>{f.title}</div>
-                <div className='text-base-content/70 mt-0.5'>{f.reason}</div>
-              </li>
-            ))}
-          </ul>
         </div>
       )}
     </div>
